rewrite viex store to module way
